BJS2-96850 delete expired ad done
