+++ src/main/java/faang/school/postservice/model/ad/Ad.java
@@ -11,8 +11,10 @@ import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
@@ -22,6 +24,8 @@ import java.time.LocalDateTime;
 @Builder
 @Entity
 @Table(name = "post_ad")
+@NoArgsConstructor
+@AllArgsConstructor
 public class Ad {
 
     @Id
