+++ src/main/java/faang/school/postservice/repository/ad/AdRepository.java
@@ -1,13 +1,13 @@
 package faang.school.postservice.repository.ad;
 
 import faang.school.postservice.model.ad.Ad;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface AdRepository extends CrudRepository<Ad, Long> {
+public interface AdRepository extends JpaRepository<Ad, Long> {
 
     @Query("SELECT a FROM Ad a WHERE a.post.id = ?1")
     Optional<Ad> findByPostId(long postId);
