+++ src/main/java/faang/school/postservice/service/ad/AdService.java
@@ -1,0 +1,6 @@
+package faang.school.postservice.service.ad;
+
+public interface AdService {
+
+    void deleteExpiredAds();
+}
